Refs:40721 Rubocop fixes and Added ConsumerRoleHelper along with Conditional check to notice attachments in notice engine
