+++ components/notifier/app/helpers/notifier/consumer_role_helper.rb
@@ -1,0 +1,119 @@
+module Notifier
+  module ConsumerRoleHelper
+    def dependent_hash(dependent, member)
+      MergeDataModels::Dependent.new(
+        {
+          first_name: dependent.first_name,
+          last_name: dependent.last_name,
+          age: dependent.age,
+          federal_tax_filing_status: filer_type(member['filer_type']),
+          expected_income_for_coverage_year: member['actual_income'].present? ? ActionController::Base.helpers.number_to_currency(member['actual_income'], :precision => 0) : "",
+          citizenship: citizen_status(member["citizen_status"]),
+          dc_resident: member['resident'],
+          tax_household_size: member['tax_hh_count'],
+          incarcerated: member['incarcerated'] == 'N' ? 'No' : 'Yes',
+          other_coverage: member["mec"].presence || 'No',
+          aptc: member['aptc'],
+          aqhp_eligible: dependent.is_aqhp_eligible,
+          uqhp_eligible: dependent.is_uqhp_eligible,
+          totally_ineligible: dependent.is_totally_ineligible
+        }
+      )
+    end
+
+    def member_hash(fam_member)
+      MergeDataModels::Dependent.new(
+        {
+          first_name: fam_member.first_name,
+          last_name: fam_member.last_name,
+          age: fam_member.age
+        }
+      )
+    end
+
+    def address_hash(mailing_address)
+      MergeDataModels::Address.new(
+        {
+          street_1: mailing_address.address_1,
+          street_2: mailing_address.address_2,
+          city: mailing_address.city,
+          state: mailing_address.state,
+          zip: mailing_address.zip
+        }
+      )
+    end
+
+    def ivl_oe_start_date_value
+      Settings.aca
+              .individual_market
+              .upcoming_open_enrollment
+              .start_on.strftime('%B %d, %Y')
+    end
+
+    def ivl_oe_end_date_value
+      Settings.aca
+              .individual_market
+              .upcoming_open_enrollment
+              .end_on.strftime('%B %d, %Y')
+    end
+
+    def expected_income_for_coverage_year_value(payload)
+      if payload['notice_params']['primary_member']['actual_income'].present?
+        ActionController::Base.helpers.number_to_currency(
+          payload['notice_params']['primary_member']['actual_income'],
+          :precision => 0
+        )
+      else
+        ""
+      end
+    end
+
+    def filer_type(type)
+      case type
+      when "Filers"
+        "Tax Filer"
+      when "Dependents"
+        "Tax Dependent"
+      when "Married Filing Jointly"
+        "Married Filing Jointly"
+      else
+        ""
+      end
+    end
+
+    def aqhp_citizen_status(status)
+      case status
+      when "US"
+        "US Citizen"
+      when "LP"
+        "Lawfully Present"
+      when "NC"
+        "US Citizen"
+      else
+        ""
+      end
+    end
+
+    def uqhp_citizen_status(status)
+      case status
+      when "us_citizen"
+        "US Citizen"
+      when "alien_lawfully_present"
+        "Lawfully Present"
+      when "indian_tribe_member"
+        "US Citizen"
+      when "lawful_permanent_resident"
+        "Lawfully Present"
+      when "naturalized_citizen"
+        "US Citizen"
+      else
+        "Ineligible Immigration Status"
+      end
+    end
+
+    def format_date(date)
+      return '' if date.blank?
+      date.strftime('%B %d, %Y')
+    end
+  end
+end
