+++ components/notifier/app/models/notifier/notice_builder.rb
@@ -143,23 +143,23 @@ module Notifier
     end
 
     def ivl_non_discrimination
-      join_pdfs [notice_path, Rails.root.join('lib/pdf_templates', ivl_non_discrimination_attachment)]
+      join_pdfs [notice_path, Rails.root.join('lib/pdf_templates', ivl_non_discrimination_attachment)] if ['projected_eligibility_notice_2', 'projected_eligibility_notice_1'].include?(event_name)
+    end
+
+    def ivl_attach_envelope
+      join_pdfs [notice_path, Rails.root.join('lib/pdf_templates', ivl_envelope_without_address)] if ['projected_eligibility_notice_2', 'projected_eligibility_notice_1'].include?(event_name)
     end
 
     def voter_application
-      join_pdfs [notice_path, Rails.root.join('lib/pdf_templates', ivl_voter_application)]
+      join_pdfs [notice_path, Rails.root.join('lib/pdf_templates', ivl_voter_application)] if ['projected_eligibility_notice_2', 'projected_eligibility_notice_1'].include?(event_name)
     end
 
     def ivl_blank_page
-      join_pdfs [notice_path, Rails.root.join('lib/pdf_templates', ivl_blank_page_attachment)]
+      join_pdfs [notice_path, Rails.root.join('lib/pdf_templates', ivl_blank_page_attachment)] if ['projected_eligibility_notice_2', 'projected_eligibility_notice_1'].include?(event_name)
     end
 
     def attach_envelope
       join_pdfs [notice_path, Rails.root.join('lib/pdf_templates', shop_envelope_without_address)]
-    end
-
-    def ivl_attach_envelope
-      join_pdfs [notice_path, Rails.root.join('lib/pdf_templates', ivl_envelope_without_address)]
     end
 
     def employee_appeal_rights
